+++ packages/adapters/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @chamba/adapters
 
+## 0.21.0
+
+### Patch Changes
+
+- @chamba/core@0.21.0
+
 ## 0.20.1
 
 ### Patch Changes